Implement suggestions after CR

+++ src/main/java/com/MyApp/budgetControl/domain/expense/dto/ExpenseRequestDTO.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import java.math.BigDecimal;
 import lombok.Value;
 
 @Value
@@ -13,7 +14,7 @@ public class ExpenseRequestDTO {
   @NotNull(message = "Cost value is mandatory")
   @DecimalMin(value = "0.01", message = "Cost value should be positive")
   @JsonProperty("expenseCost")
-  private final double expenseCost;
+  private final BigDecimal expenseCost;
 
   @NotBlank(message = "Category is mandatory")
   @Size(max = 36)
