+++ src/main/java/com/MyApp/budgetControl/report/dto/MonthlyExpenseReportDTO.java
@@ -1,6 +1,7 @@
 package com.MyApp.budgetControl.report.dto;
 
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 import lombok.Value;
 
 @Value
@@ -8,6 +9,6 @@ public class MonthlyExpenseReportDTO {
 
   @Id
   private final String userName;
-  private final double monthlyExpensesTotal;
+  private final BigDecimal monthlyExpensesTotal;
 }
